only delete old events, if they really container the sync-category as it seems the search does not correctly return only events with the sync-category

+++ api/src/CalDAV/Sync.php
@@ -496,7 +496,10 @@ EOT, $this->header([
 		// delete NOT imported $old_events
 		foreach($old_events as $old_event)
 		{
-			$ical_class->delete($old_event['id'], 0, true, true);
+			if (!empty($old_event['category']) && strstr(','.$old_event['category'].',', ','.$modifications['cat_id'].',') !== false)
+			{
+				$ical_class->delete($old_event['id'], 0, true, true);
+			}
 		}
 	}
 
@@ -563,7 +566,10 @@ EOT, $this->header([
 		// delete NOT imported $old_events
 		foreach($old_events as $old_event)
 		{
-			$ical_class->delete($old_event['id'], 0, true, true);
+			if (!empty($old_event['category']) && strstr(','.$old_event['category'].',', ','.$modifications['cat_id'].',') !== false)
+			{
+				$ical_class->delete($old_event['id'], 0, true, true);
+			}
 		}
 		$etag = $response_header['etag'];
 	}
